+++ assets/javascripts/kitten/components/accordeon/components/context.js
@@ -1,8 +1,7 @@
-import React from 'react'
+import { createContext } from 'react'
 
-export const Context = React.createContext({
+export const Context = createContext({
   selectedItem: null,
   updateSelectedItem: () => {},
-  isAnimated: true,
   componentId: 'accordeon',
 })