+++ assets/javascripts/kitten/components/form/radio-button/index.js
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import styled from 'styled-components'
@@ -7,88 +7,110 @@ import COLORS from '../../../constants/colors-config'
 import TYPOGRAPHY from '../../../constants/typography-config'
 import { ScreenConfig } from '../../../constants/screen-config'
 
-const StyledRadioButton = styled.label`
-  display: inline-flex;
-  align-items: center;
-  align-self: flex-start;
-  flex-wrap: wrap;
-  cursor: pointer;
+const StyledRadioButton = styled.div`
+  display: inline;
 
   &, & * {
     transition: all .2s ease;
   }
 
-  .k-Form-RadioButton__input {
-    appearance: none;
+  /* label */
 
-    box-sizing: border-box;
-    display: inline-block;
-    flex-shrink: 0;
-    width: ${pxToRem(16)};
-    height: ${pxToRem(16)};
-    margin: ${pxToRem(2)} ${pxToRem(10)} 0 0;
+  .k-Form-RadioButton__label {
+    display: flex;
+    align-items: center;
+    align-self: flex-start;
+    flex-wrap: wrap;
+    cursor: pointer;
 
-    content: '';
-    background-color: ${COLORS.background1};
-    border: ${pxToRem(2)} solid ${COLORS.line1};
-    border-radius: 50%;
 
-    transition: border .1s;
+    &::before {
+      box-sizing: border-box;
+      display: block;
+      flex-shrink: 0;
+      width: ${pxToRem(16)};
+      height: ${pxToRem(16)};
+      margin: ${pxToRem(2)} ${pxToRem(10)} 0 0;
 
-    &:focus {
-      outline: none;
-      border-color: ${COLORS.primary1};
-    }
+      content: '';
+      background-position: center;
+      background-repeat: no-repeat;
+      background-color: ${COLORS.background1};
+      border: ${pxToRem(2)} solid ${COLORS.line1};
+      border-radius: 50%;
 
-    &:checked {
-      position: relative;
-      border: ${pxToRem(5)} solid ${COLORS.primary1};
+      transition: border .2s, background .2s;
     }
   }
 
-  &.k-Form-RadioButton--error .k-Form-RadioButton__input {
-    border-color: ${COLORS.error3};
+  /* Hide input */
 
-    &:checked {
+  .k-Form-RadioButton__input {
+    position: absolute;
+    opacity: 0;
+    width: 0;
+    height: 0;
+    overflow: hidden;
+  }
+
+  /* Label statuses */
+
+  .k-Form-RadioButton__input:focus + .k-Form-RadioButton__label::before {
+    outline: none;
+    border-color: ${COLORS.primary1};
+  }
+  .k-Form-RadioButton__input:checked + .k-Form-RadioButton__label::before {
+    border: ${pxToRem(5)} solid ${COLORS.primary1};
+  }
+
+  /* Error */
+
+  &.k-Form-RadioButton--error {
+    .k-Form-RadioButton__label::before {
+      border-color: ${COLORS.error3};
+    }
+    .k-Form-RadioButton__input:checked:not(:disabled) + .k-Form-RadioButton__label::before {
       border-color: ${COLORS.error};
     }
   }
 
-  & .k-Form-RadioButton__input:disabled {
-    background-color: ${COLORS.line1};
-    border-color: ${COLORS.line2};
+  /* Disabled */
 
-    &:checked {
+  .k-Form-RadioButton__input:disabled {
+    & + .k-Form-RadioButton__label {
+      border-color: ${COLORS.line2};
+    }
+
+    & + .k-Form-RadioButton__label::before {
+      background-color: ${COLORS.line1};
+      border-color: ${COLORS.line2};
+
+    }
+    &:checked + .k-Form-RadioButton__label::before {
       background-color: ${COLORS.background1};
       border-color: ${COLORS.line2};
     }
   }
 
-  .k-Form-RadioButton__label {
+
+  /* Label text styles */
+
+  .k-Form-RadioButton__labelText {
     ${TYPOGRAPHY.fontStyles.light}
     font-size: ${stepToRem(-1)};
-
-    .k-Form-RadioButton__input:disabled + & {
-      cursor: not-allowed;
-    }
-
-    .k-Form-RadioButton--error & {
-      border-color: ${COLORS.error3};
-    }
   }
 
-  &:active .k-Form-RadioButton__label,
-  .k-Form-RadioButton__label:active,
-  .k-Form-RadioButton__input:active + .k-Form-RadioButton__label {
+  :active .k-Form-RadioButton__labelText,
+  .k-Form-RadioButton__input:active + .k-Form-RadioButton__labelText {
     color: ${COLORS.primary3};
   }
 
-  .k-Form-RadioButton__label--large {
+  .k-Form-RadioButton__labelText--large {
     ${TYPOGRAPHY.fontStyles.regular}
     font-size: ${stepToRem(2)};
   }
 
-  .k-Form-RadioButton__label--withContents {
+  .k-Form-RadioButton__labelText--withContents {
     ${TYPOGRAPHY.fontStyles.regular}
   }
 
@@ -108,51 +130,75 @@ const StyledRadioButton = styled.label`
     font-size: ${stepToRem(-1)};
   }
 
-  &.k-Form-RadioButton--orion {
-    box-sizing: border-box;
-    padding: ${pxToRem(15)};
-    height: ${pxToRem(60)};
-    border-radius: ${pxToRem(6)};
-    border: 2px solid ${COLORS.line1};
 
-    &:active,
-    &:focus-within,
-    &:focus,
-    &.k-Form-RadioButton--isChecked {
-      border-color: ${COLORS.primary1};
+  /* VARIANT = "orion" */
+
+  &.k-Form-RadioButton--orion {
+
+    .k-Form-RadioButton__label {
+      box-sizing: border-box;
+      padding: ${pxToRem(15)};
+      height: ${pxToRem(60)};
+      border-radius: ${pxToRem(6)};
+      border: 2px solid ${COLORS.line1};
+    }
+
+    .k-Form-RadioButton__input:checked:not(:disabled),
+    .k-Form-RadioButton__input:focus:not(:disabled) {
+      & + .k-Form-RadioButton__label {
+        border-color: ${COLORS.primary1};
+      }
     }
 
     &.k-Form-RadioButton--error {
-      border-color: ${COLORS.error3};
+      .k-Form-RadioButton__input:not(:disabled) + .k-Form-RadioButton__label {
+        border-color: ${COLORS.error3};
+      }
+    }
+
+    .k-Form-RadioButton__labelText {
+      ${TYPOGRAPHY.fontStyles.regular};
     }
 
     @media (min-width: ${ScreenConfig.S.min}px) {
-      height: ${pxToRem(80)};
-      border-radius: ${pxToRem(8)};
-
-      .k-Form-RadioButton__input {
-        width: ${pxToRem(20)};
-        height: ${pxToRem(20)};
-
-        &:checked {
-          border-width: ${pxToRem(6)};
-        }
+      .k-Form-RadioButton__label {
+        height: ${pxToRem(80)};
+        border-radius: ${pxToRem(8)};
       }
 
-      .k-Form-RadioButton__label {
+      .k-Form-RadioButton__label::before {
+        width: ${pxToRem(20)};
+        height: ${pxToRem(20)};
+      }
+
+      .k-Form-RadioButton__labelText {
         font-size: ${stepToRem(0)};
       }
 
-      .k-Form-RadioButton__input:disabled + .k-Form-RadioButton__label {
-        color: ${COLORS.font2};
+      .k-Form-RadioButton__input:checked + .k-Form-RadioButton__label::before {
+        border-width: ${pxToRem(6)};
       }
-
-    }
-
-    .k-Form-RadioButton__label {
-      ${TYPOGRAPHY.fontStyles.regular};
     }
   }
+
+
+  /* DESIGN = "check" */
+
+  &.k-Form-RadioButton--check {
+    .k-Form-RadioButton__input:checked + .k-Form-RadioButton__label::before {
+      border-width: ${pxToRem(2)};
+      background-color: ${COLORS.primary1};
+      background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 11 8' width='11' height='8' fill='%23fff' %3E%3Cpath d='M.904 4.06l1.414-1.414L4.44 4.768 8.682.525l1.414 1.414L4.44 7.596.904 4.061z' /%3E%3C/svg%3E");
+    }
+
+    &.k-Form-RadioButton--error .k-Form-RadioButton__input:checked:not(:disabled) + .k-Form-RadioButton__label::before {
+      background-color: ${COLORS.error};
+    }
+    .k-Form-RadioButton__input:disabled + .k-Form-RadioButton__label::before {
+      background-color: ${COLORS.line2};
+    }
+  }
+
 `
 
 export const RadioButton = ({
@@ -166,60 +212,49 @@ export const RadioButton = ({
   error,
   disabled,
   variant,
-  onChange,
+  design,
   ...inputProps
 }) => {
-  const inputEl = useRef(null)
-  const [isChecked, setChecked] = useState(false)
-
-  const handleChange = changeEvent => {
-    setChecked(inputEl.current.checked)
-
-    onChange(changeEvent)
-  }
-
   return (
     <StyledRadioButton
-      htmlFor={id}
       className={classNames(
         'k-Form-RadioButton',
         className,
         `k-Form-RadioButton--${variant}`,
+        `k-Form-RadioButton--${design}`,
         {
           'k-Form-RadioButton--error': error,
-          'k-Form-RadioButton--isChecked': isChecked,
         },
       )}
     >
       <input
-        ref={inputEl}
         id={id}
         type="radio"
         className={classNames('k-Form-RadioButton__input', inputClassName)}
         disabled={disabled}
-        onChange={handleChange}
         {...inputProps}
       />
-
-      <span
-        className={classNames(
-          'k-Form-RadioButton__label',
-          { 'k-Form-RadioButton__label--large': large },
-          { 'k-Form-RadioButton__label--withContents': !!children },
-        )}
-      >
-        {text}
-      </span>
-
-      {children && (
-        <div
-          className={classNames('k-Form-RadioButton__labelContents', {
-            'k-Form-RadioButton__labelContents--large': largeContent,
-          })}
+      <label htmlFor={id} className="k-Form-RadioButton__label">
+        <span
+          className={classNames(
+            'k-Form-RadioButton__labelText',
+            { 'k-Form-RadioButton__labelText--large': large },
+            { 'k-Form-RadioButton__labelText--withContents': !!children },
+          )}
         >
-          {children}
-        </div>
-      )}
+          {text}
+        </span>
+
+        {children && (
+          <div
+            className={classNames('k-Form-RadioButton__labelContents', {
+              'k-Form-RadioButton__labelContents--large': largeContent,
+            })}
+          >
+            {children}
+          </div>
+        )}
+      </label>
     </StyledRadioButton>
   )
 }
@@ -232,7 +267,7 @@ RadioButton.propTypes = {
   error: PropTypes.bool,
   disabled: PropTypes.bool,
   variant: PropTypes.oneOf(['andromeda', 'orion']),
-  onChange: PropTypes.func,
+  design: PropTypes.oneOf(['disc', 'check']),
 }
 
 RadioButton.defaultProps = {
@@ -241,5 +276,5 @@ RadioButton.defaultProps = {
   error: false,
   disabled: false,
   variant: 'andromeda',
-  onChange: () => {},
+  design: 'disc',
 }